requirements.txt
apex==0.9.10.dev0
numpy==1.19.5
Cython==0.29.23
opencv-python>=4.2.0.32
scikit-learn>=0.24.1
scikit-image>=0.16.2
numba==0.53.1
Pillow==8.2.0
loguru==0.5.3
pycocotools==2.0.0
matplotlib==3.3.4
vidgear==0.1.7
Shapely==1.7.1
pandas==1.1.5
geopandas==0.9.0
scipy>=1.4.1
filterpy==1.4.5
lap==0.4.0
imantics==0.1.12
beautifulsoup4==4.9.3
yacs==0.1.8
psutil==5.8.0
six==1.15.0
astunparse
future
pyyaml
setuptools
setuptools_rust
typing_extensions

tensorflow-gpu==2.6.0
keras==2.6.0
tensorrt==7.2.3.4
onnx==1.4.1
onnxruntime==1.8.1
pycuda==2021.1
tf2onnx==1.9.1

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0

--find-links https://download.openmmlab.com/mmcv/dist/cu111/torch1.9.0/index.html
mmcv-full==1.3.14
